refator: refactor dpos actions
