+++ Lib9c/Action/DPoS/Control/ValidatorRewardsCtrl.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using Libplanet.Action.State;
 using Libplanet.Crypto;
 using Libplanet.Types.Assets;
+using Nekoyume.Action.DPoS.Misc;
 using Nekoyume.Action.DPoS.Model;
+using Nekoyume.Action.DPoS.Util;
 
 namespace Nekoyume.Action.DPoS.Control
 {
