+++ Lib9c/Action/DPoS/Control/ValidatorSetCtrl.cs
@@ -6,6 +6,7 @@ using Libplanet.Crypto;
 using Nekoyume.Action.DPoS.Exception;
 using Nekoyume.Action.DPoS.Misc;
 using Nekoyume.Action.DPoS.Model;
+using Nekoyume.Action.DPoS.Util;
 using Nekoyume.Module;
 
 namespace Nekoyume.Action.DPoS.Control
