+++ Lib9c/Action/DPoS/Delegate.cs
@@ -5,48 +5,42 @@ using Libplanet.Action.State;
 using Libplanet.Crypto;
 using Libplanet.Types.Assets;
 using Nekoyume.Action.DPoS.Control;
-using Nekoyume.Action.DPoS.Exception;
 using Nekoyume.Action.DPoS.Misc;
 using Nekoyume.Action.DPoS.Util;
 
-namespace Nekoyume.Action.DPoS.Sys
+namespace Nekoyume.Action.DPoS
 {
     /// <summary>
-    /// A system action for DPoS that promotes non-validator node to a validator.
+    /// A system action for DPoS that <see cref="Delegate"/> specified <see cref="Amount"/>
+    /// of tokens to a given <see cref="Validator"/>.
     /// </summary>
     [ActionType(ActionTypeValue)]
-    public sealed class PromoteValidator : ActionBase
+    public sealed class Delegate : ActionBase
     {
-        private const string ActionTypeValue = "promote_validator";
+        private const string ActionTypeValue = "delegate";
 
         /// <summary>
-        /// Create a new instance of <see cref="PromoteValidator"/> action.
+        /// Creates a new instance of <see cref="Delegate"/> action.
         /// </summary>
-        /// <param name="validator">The <see cref="PublicKey"/> of the target
-        /// to promote validator.</param>
-        /// <param name="amount">The amount of the asset to be initialize delegation.</param>
-        public PromoteValidator(PublicKey validator, FungibleAssetValue amount)
+        /// <param name="validator">The <see cref="Address"/> of the validator
+        /// to delegate tokens.</param>
+        /// <param name="amount">The amount of the asset to be delegated.</param>
+        public Delegate(Address validator, FungibleAssetValue amount)
         {
             Validator = validator;
             Amount = amount;
         }
 
-        public PromoteValidator()
+        public Delegate()
         {
             // Used only for deserialization.  See also class Libplanet.Action.Sys.Registry.
-            // FIXME: do not fill ambiguous validator field.
-            // Suggestion: https://gist.github.com/riemannulus/7405e0d361364c6afa0ab433905ae81c
-            Validator = new PrivateKey().PublicKey;
         }
 
         /// <summary>
-        /// The <see cref="PublicKey"/> of the target promoting to a validator.
+        /// The <see cref="Address"/> of the validator to <see cref="Delegate"/>.
         /// </summary>
-        public PublicKey Validator { get; set; }
+        public Address Validator { get; set; }
 
-        /// <summary>
-        /// The amount of the asset to be initially delegated.
-        /// </summary>
         public FungibleAssetValue Amount { get; set; }
 
         /// <inheritdoc cref="IAction.PlainValue"/>
@@ -59,7 +53,7 @@ namespace Nekoyume.Action.DPoS.Sys
         public override void LoadPlainValue(IValue plainValue)
         {
             var dict = (Bencodex.Types.Dictionary)plainValue;
-            Validator = dict["validator"].ToPublicKey();
+            Validator = dict["validator"].ToAddress();
             Amount = dict["amount"].ToFungibleAssetValue();
         }
 
@@ -67,16 +61,13 @@ namespace Nekoyume.Action.DPoS.Sys
         public override IWorld Execute(IActionContext context)
         {
             IActionContext ctx = context;
-            if (!ctx.Signer.Equals(Validator.Address))
-            {
-                throw new PublicKeyAddressMatchingException(ctx.Signer, Validator);
-            }
-
             var states = ctx.PreviousState;
+
+            // if (ctx.Rehearsal)
+            // Rehearsal mode is not implemented
             var nativeTokens = ImmutableHashSet.Create(
                 Asset.GovernanceToken, Asset.ConsensusToken, Asset.Share);
-
-            states = ValidatorCtrl.Create(
+            states = DelegateCtrl.Execute(
                 states,
                 ctx,
                 ctx.Signer,
