+++ Lib9c/Action/DPoS/PromoteValidator.cs
@@ -5,46 +5,47 @@ using Libplanet.Action.State;
 using Libplanet.Crypto;
 using Libplanet.Types.Assets;
 using Nekoyume.Action.DPoS.Control;
+using Nekoyume.Action.DPoS.Exception;
 using Nekoyume.Action.DPoS.Misc;
-using Nekoyume.Action.DPoS.Model;
 using Nekoyume.Action.DPoS.Util;
 
-namespace Nekoyume.Action.DPoS.Sys
+namespace Nekoyume.Action.DPoS
 {
     /// <summary>
-    /// A system action for DPoS that cancel <see cref="Undelegate"/> specified
-    /// <see cref="Amount"/> of tokens to a given <see cref="Validator"/>.
+    /// A system action for DPoS that promotes non-validator node to a validator.
     /// </summary>
     [ActionType(ActionTypeValue)]
-    public sealed class CancelUndelegation : ActionBase
+    public sealed class PromoteValidator : ActionBase
     {
-        private const string ActionTypeValue = "cancel_undelegation";
+        private const string ActionTypeValue = "promote_validator";
 
         /// <summary>
-        /// Creates a new instance of <see cref="CancelUndelegation"/> action.
+        /// Create a new instance of <see cref="PromoteValidator"/> action.
         /// </summary>
-        /// <param name="validator">The <see cref="amount"/> of the validator
-        /// to delegate tokens.</param>
-        /// <param name="amount">The amount of the asset to be delegated.</param>
-        public CancelUndelegation(Address validator, FungibleAssetValue amount)
+        /// <param name="validator">The <see cref="PublicKey"/> of the target
+        /// to promote validator.</param>
+        /// <param name="amount">The amount of the asset to be initialize delegation.</param>
+        public PromoteValidator(PublicKey validator, FungibleAssetValue amount)
         {
             Validator = validator;
             Amount = amount;
         }
 
-        public CancelUndelegation()
+        public PromoteValidator()
         {
             // Used only for deserialization.  See also class Libplanet.Action.Sys.Registry.
+            // FIXME: do not fill ambiguous validator field.
+            // Suggestion: https://gist.github.com/riemannulus/7405e0d361364c6afa0ab433905ae81c
+            Validator = new PrivateKey().PublicKey;
         }
 
         /// <summary>
-        /// The <see cref="Address"/> of the validator
-        /// to cancel the <see cref="Undelegate"/> and <see cref="Delegate"/>.
+        /// The <see cref="PublicKey"/> of the target promoting to a validator.
         /// </summary>
-        public Address Validator { get; set; }
+        public PublicKey Validator { get; set; }
 
         /// <summary>
-        /// The amount of the asset to be delegated.
+        /// The amount of the asset to be initially delegated.
         /// </summary>
         public FungibleAssetValue Amount { get; set; }
 
@@ -58,7 +59,7 @@ namespace Nekoyume.Action.DPoS.Sys
         public override void LoadPlainValue(IValue plainValue)
         {
             var dict = (Bencodex.Types.Dictionary)plainValue;
-            Validator = dict["validator"].ToAddress();
+            Validator = dict["validator"].ToPublicKey();
             Amount = dict["amount"].ToFungibleAssetValue();
         }
 
@@ -66,16 +67,20 @@ namespace Nekoyume.Action.DPoS.Sys
         public override IWorld Execute(IActionContext context)
         {
             IActionContext ctx = context;
+            if (!ctx.Signer.Equals(Validator.Address))
+            {
+                throw new PublicKeyAddressMatchingException(ctx.Signer, Validator);
+            }
+
             var states = ctx.PreviousState;
             var nativeTokens = ImmutableHashSet.Create(
                 Asset.GovernanceToken, Asset.ConsensusToken, Asset.Share);
 
-            // if (ctx.Rehearsal)
-            // Rehearsal mode is not implemented
-            states = UndelegateCtrl.Cancel(
+            states = ValidatorCtrl.Create(
                 states,
                 ctx,
-                Undelegation.DeriveAddress(ctx.Signer, Validator),
+                ctx.Signer,
+                Validator,
                 Amount,
                 nativeTokens);
