+++ Lib9c/Action/DPoS/Sys/UpdateValidators.cs
@@ -1,23 +1,22 @@
-using System.Collections.Immutable;
 using Bencodex.Types;
 using Libplanet.Action;
 using Libplanet.Action.State;
 using Nekoyume.Action.DPoS.Control;
-using Nekoyume.Action.DPoS.Misc;
 using Nekoyume.Action.DPoS.Model;
 using Nekoyume.Module;
 
-namespace Nekoyume.Action.DPoS
+namespace Nekoyume.Action.DPoS.Sys
 {
     /// <summary>
-    /// A EndBlock action for DPoS that updates <see cref="ValidatorSet"/>.
+    /// An action for update validators.
+    /// Should be executed at the end of the block.
     /// </summary>
-    public sealed class DPoSEndBlockAction : ActionBase
+    public sealed class UpdateValidators : ActionBase
     {
         /// <summary>
-        /// Creates a new instance of <see cref="DPoSBeginBlockAction"/>.
+        /// Creates a new instance of <see cref="AllocateReward"/>.
         /// </summary>
-        public DPoSEndBlockAction()
+        public UpdateValidators()
         {
         }
 
@@ -34,8 +33,6 @@ namespace Nekoyume.Action.DPoS
         public override IWorld Execute(IActionContext context)
         {
             var states = context.PreviousState;
-
-            // Update ValidatorSet
             states = ValidatorSetCtrl.Update(states, context);
             ValidatorSet bondedSet;
             (states, bondedSet) = ValidatorSetCtrl.FetchBondedValidatorSet(states);
