+++ Lib9c/Action/DPoS/WithdrawDelegator.cs
@@ -6,58 +6,49 @@ using Libplanet.Crypto;
 using Libplanet.Types.Assets;
 using Nekoyume.Action.DPoS.Control;
 using Nekoyume.Action.DPoS.Misc;
+using Nekoyume.Action.DPoS.Model;
 using Nekoyume.Action.DPoS.Util;
+using Nekoyume.Module;
 
-namespace Nekoyume.Action.DPoS.Sys
+namespace Nekoyume.Action.DPoS
 {
     /// <summary>
-    /// A system action for DPoS that cancels <see cref="Delegate"/> specified
-    /// <see cref="ShareAmount"/> of shared tokens to a given <see cref="Validator"/>.
+    /// A system action for DPoS that withdraws reward tokens from given <see cref="Validator"/>.
     /// </summary>
     [ActionType(ActionTypeValue)]
-    public sealed class Undelegate : ActionBase
+    public sealed class WithdrawDelegator : ActionBase
     {
-        private const string ActionTypeValue = "undelegate";
+        private const string ActionTypeValue = "withdraw_delegator";
 
         /// <summary>
-        /// Creates a new instance of <see cref="Undelegate"/> action.
+        /// Creates a new instance of <see cref="WithdrawDelegator"/> action.
         /// </summary>
-        /// <param name="validator">The <see cref="amount"/> of the validator
-        /// to undelegate tokens.</param>
-        /// <param name="amount">The amount of the asset to be undelegated.</param>
-        public Undelegate(Address validator, FungibleAssetValue amount)
+        /// <param name="validator">The <see cref="Address"/> of the validator
+        /// from which to withdraw the tokens.</param>
+        public WithdrawDelegator(Address validator)
         {
             Validator = validator;
-            ShareAmount = amount;
         }
 
-        internal Undelegate()
+        public WithdrawDelegator()
         {
             // Used only for deserialization.  See also class Libplanet.Action.Sys.Registry.
         }
 
         /// <summary>
-        /// The <see cref="Address"/> of the validator to cancel the <see cref="Delegate"/>.
+        /// The <see cref="Address"/> of the validator to withdraw.
         /// </summary>
         public Address Validator { get; set; }
-
-        /// <summary>
-        /// The amount of the asset to be undelegated.
-        /// </summary>
-        public FungibleAssetValue ShareAmount { get; set; }
 
         /// <inheritdoc cref="IAction.PlainValue"/>
         public override IValue PlainValue => Bencodex.Types.Dictionary.Empty
             .Add("type_id", new Text(ActionTypeValue))
-            .Add("validator", Validator.Serialize())
-            .Add("amount", ShareAmount.Serialize());
+            .Add("validator", Validator.Serialize());
 
         /// <inheritdoc cref="IAction.LoadPlainValue(IValue)"/>
         public override void LoadPlainValue(IValue plainValue)
         {
-            var dict = (Bencodex.Types.Dictionary)plainValue;
-            Validator = dict["validator"].ToAddress();
-            ShareAmount = dict["amount"].ToFungibleAssetValue();
+            Validator = plainValue.ToAddress();
         }
 
         /// <inheritdoc cref="IAction.Execute(IActionContext)"/>
@@ -70,13 +61,27 @@ namespace Nekoyume.Action.DPoS.Sys
 
             // if (ctx.Rehearsal)
             // Rehearsal mode is not implemented
-            states = UndelegateCtrl.Execute(
+            states = DelegateCtrl.Distribute(
                 states,
                 ctx,
-                ctx.Signer,
-                Validator,
-                ShareAmount,
-                nativeTokens);
+                nativeTokens,
+                Delegation.DeriveAddress(ctx.Signer, Validator));
+
+#pragma warning disable LAA1002
+            foreach (Currency nativeToken in nativeTokens)
+            {
+                FungibleAssetValue reward = states.GetBalance(
+                    AllocateReward.RewardAddress(ctx.Signer), nativeToken);
+                if (reward.Sign > 0)
+                {
+                    states = states.TransferAsset(
+                        ctx,
+                        AllocateReward.RewardAddress(ctx.Signer),
+                        ctx.Signer,
+                        reward);
+                }
+            }
+#pragma warning restore LAA1002
 
             return states;
         }
